pyproject.toml
[tool.poetry]
name = "pygmh-mrb"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
repository = "https://github.com/FintelmannLabDevelopmentTeam/pygmh-mrb/"

[tool.poetry.dependencies]
pygmh = { git = "https://github.com/FintelmannLabDevelopmentTeam/pygmh.git", branch = "master" }
pynrrd = "^0.4.2"
python = "^3.7"
numpy = "^1.18"

[tool.poetry.dev-dependencies]
pytest = "^5.4.1"

[tool.poetry.urls]
issues = "https://github.com/FintelmannLabDevelopmentTeam/pygmh-mrb/issues"
